requirements.txt
asgiref==3.8.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-resized==1.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyTelegramBotAPI==4.15.4
python-dotenv==1.1.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.3
